Use fixed-width count in TextEditor events C API
te_change_batch.count and SetAutoCompleteSuggestions' count were size_t,
so cinterop exposed UInt on androidNativeArm32 and ULong elsewhere. The
expect/actual type-consistency checks reject platform-dependent widths in
inferred signatures during publish, and the arm32 native compile failed.
Make both counts uint32_t for a stable cross-platform cinterop surface
(the public publish/compile now passes on all androidNative ABIs).

+++ jni/c_api/text_editor_events_c.cc
@@ -103,7 +103,7 @@ void te_change_batch_free(te_change_batch* batch) {
 // Async autocomplete
 // =========================================================================
 
-void te_set_auto_complete_suggestions(te_editor* editor, const char* const* values, size_t count) {
+void te_set_auto_complete_suggestions(te_editor* editor, const char* const* values, uint32_t count) {
     TextEditor* te = reinterpret_cast<TextEditor*>(editor);
     std::vector<std::string> suggestions;
     suggestions.reserve(count);
